requirements.txt
# Core dependencies - let resolver handle versions
dbt-snowflake
snowflake-connector-python
pandas
datasets
huggingface_hub
requests
boto3
anthropic
